+++ test/action_client_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class ActionClient::Test < ActiveSupport::TestCase
-  test "truth" do
-    assert_kind_of Module, ActionClient
-  end
-end